Hurst zones and scan list modified database and firmware updated
Updated to the latest firmware and CPS 4.01.006. If you want to use this CP, you MUST upgrade the firmware on your radio to vs. 3.02.00+, Bootloader vs. R2.04S AND front panel firmware vs. 2.00.08+ in order to use this codeplug.

Upgrade files and instructions can be found on the Connect Systems webpage: http://www.connectsystems.com/software/top/CS800.htm

Follow them very carefully. This is an extensive upgrade process, and takes time, but it doubles the contacts database limits from 64,000 to 130,000 contacts.