Initial submission by Noel, KF5SLK
Codeplug for Baofeng DM-1701 dual band DMR radio built using Baofeng CPS programmer V1.05.

Geared for DFW area, but all of the available repeaters are available for future Zone builds.

Includes a VHF and UHF repeater zone for DFW and a Simplex zone for commonly used simplex frequencies. Not recommended for a primary V/U radio other than DMR as the squelch for this radio is basically on/off and it picked up a lot of noise in the testing I did.